+++ crates/api-desc/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Minor
 
+- Add `platform::update` module
 - Add `platform` module with `platform::reboot()` function
 - Call some `env_dispatch` function instead of panicking in `native`
 - Prefix function symbols by `env_` with `native` feature
@@ -76,4 +77,4 @@
 
 ## 0.1.0
 
-<!-- Increment to skip CHANGELOG.md test: 10 -->
+<!-- Increment to skip CHANGELOG.md test: 11 -->